api/requirements.txt
Flask
kafka-python
pymysql
python-dotenv
